Projects container added. Profile replaced with account. Actions removed.
