+++ app/containers/Account/Account.scss
@@ -1,0 +1,3 @@
+.Account {
+
+}
