--- app/containers/Cars/Cars.scss
@@ -1,3 +1,0 @@
-.Cars {
-
-}
