--- app/containers/Profile/Profile.scss
@@ -1,21 +1,0 @@
-.Profile {
-  width: 100%;
-  display: flex;
-  flex-direction:column;
-  align-items: center;
-  justify-content: center;
-  &-Data {
-    display: flex;
-    flex-direction:column;
-  }
-  &-Datapoint {
-    margin-bottom: 1.2rem;
-  }
-  &-Username {
-    font-size: 2rem;
-  }
-  &-Email {
-    margin-bottom: 2.5rem;
-
-  }
-}
