+++ app/containers/Projects/Projects.scss
@@ -1,0 +1,3 @@
+.Projects {
+
+}
