+++ app/containers/index.js
@@ -1,9 +1,8 @@
 export App from './App/App';
+export Account from './Account/Account';
 export Home from './Home/Home';
 export About from './About/About';
 export Login from './Login/Login';
 export Signup from './Signup/Signup';
-export Profile from './Profile/Profile';
-export Cars from './Cars/Cars';
 export RequireLogin from './RequireLogin/RequireLogin';
 export NotFound from './NotFound/NotFound';
